requirements.txt
alembic==1.4.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
dataset==1.3.0
Flask==1.1.1
Flask-Login==0.5.0
gunicorn==20.0.4
htmlmin==0.1.12
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
neo4j==1.7.6
neobolt==1.7.17
neotime==1.7.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.15
urllib3==1.25.8
Werkzeug==1.0.0
